+++ stream-access-control/access-control-core/src/main/java/com/backbase/stream/configuration/AccessControlConfiguration.java
@@ -14,6 +14,7 @@ import com.backbase.dbs.accesscontrol.api.service.v2.UsersApi;
 import com.backbase.dbs.user.api.service.v2.IdentityManagementApi;
 import com.backbase.dbs.user.api.service.v2.UserManagementApi;
 import com.backbase.dbs.user.profile.api.service.v2.UserProfileManagementApi;
+import com.backbase.identity.integration.api.service.v1.IdentityIntegrationServiceApi;
 import com.backbase.stream.config.BackbaseStreamConfigurationProperties;
 import com.backbase.stream.product.configuration.ProductConfiguration;
 import com.backbase.stream.product.service.ArrangementService;
@@ -24,6 +25,9 @@ import com.backbase.stream.service.UserProfileService;
 import com.backbase.stream.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.text.DateFormat;
+import java.util.Optional;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -36,6 +40,7 @@ import org.springframework.web.reactive.function.client.WebClient;
 @Configuration
 @EnableConfigurationProperties(BackbaseStreamConfigurationProperties.class)
 @Import(ProductConfiguration.class)
+@Slf4j
 public class AccessControlConfiguration {
 
     private final BackbaseStreamConfigurationProperties backbaseStreamConfigurationProperties;
@@ -62,10 +67,13 @@ public class AccessControlConfiguration {
     }
 
     @Bean
-    public UserService userService(com.backbase.dbs.user.api.service.ApiClient usersApiClient) {
+    public UserService userService(com.backbase.dbs.user.api.service.ApiClient usersApiClient,
+                                   Optional<com.backbase.identity.integration.api.service.ApiClient>
+                                       identityApiClient) {
         UserManagementApi usersApi = new UserManagementApi(usersApiClient);
         IdentityManagementApi identityManagementApi = new IdentityManagementApi(usersApiClient);
-        return new UserService(usersApi, identityManagementApi);
+        return new UserService(usersApi, identityManagementApi,
+            identityApiClient.map(IdentityIntegrationServiceApi::new));
     }
 
     @Bean
@@ -125,6 +133,24 @@ public class AccessControlConfiguration {
             new com.backbase.dbs.accesscontrol.api.service.ApiClient(
                 dbsWebClient, objectMapper, dateFormat);
         apiClient.setBasePath(backbaseStreamConfigurationProperties.getDbs().getAccessControlBaseUrl());
+        return apiClient;
+    }
+
+    @Bean
+    @ConditionalOnProperty(value = "backbase.stream.legalentity.sink.use-identity-integration")
+    protected com.backbase.identity.integration.api.service.ApiClient identityApiClient(
+        WebClient dbsWebClient,
+        ObjectMapper objectMapper,
+        DateFormat dateFormat) {
+        if (backbaseStreamConfigurationProperties.getIdentity() == null
+            || backbaseStreamConfigurationProperties.getIdentity().getIdentityIntegrationBaseUrl() == null) {
+            log.error("missing identity url configuration");
+            return null;
+        }
+        com.backbase.identity.integration.api.service.ApiClient apiClient =
+            new com.backbase.identity.integration.api.service.ApiClient(
+                dbsWebClient, objectMapper, dateFormat);
+        apiClient.setBasePath(backbaseStreamConfigurationProperties.getIdentity().getIdentityIntegrationBaseUrl());
         return apiClient;
     }