+++ stream-legal-entity/legal-entity-core/src/main/java/com/backbase/stream/LegalEntitySaga.java
@@ -524,11 +524,18 @@ public class LegalEntitySaga implements StreamTaskExecutor<LegalEntityTask> {
                 .switchIfEmpty(Mono.error(new StreamTaskException(streamTask, "Realm: " + legalEntity.getRealmName() + " not found!")))
                 .then(userService.linkLegalEntityToRealm(legalEntity))
                 .then(userService.createOrImportIdentityUser(user, legalEntity.getInternalId()))
+                .flatMap(u -> updateUserStatus(u, legalEntity.getRealmName()))
                 .doOnNext(existingUser -> {
                     user.setInternalId(existingUser.getInternalId());
                     streamTask.info(IDENTITY_USER, CREATED, user.getExternalId(), user.getInternalId(), "User %s created", existingUser.getExternalId());
                 });
         return getExistingIdentityUser.switchIfEmpty(createNewIdentityUser);
+    }
+
+    private Mono<User> updateUserStatus(User user, String realm) {
+        log.info("changing user {} status to locked {}", user.getInternalId(), user.getLocked());
+        return userService.changeEnableStatus(user, realm)
+            .thenReturn(user);
     }
 
     private Mono<LegalEntityTask> setupServiceAgreement(LegalEntityTask streamTask) {